Item13626: More foswikification
Not feeling very good about this one.  I don't have any way to test. And
I think some parts of the implementation are very wrong.  it's
establishing the user during "new" processing ... I think it should be
in getUser

Testing with a bogus sso server, it loops and apache has to be
restarted.